+++ src/FacebookConnector.php
@@ -10,13 +10,16 @@
 
 namespace itscoding\facebookconnector;
 
+use craft\web\twig\variables\CraftVariable;
+use itscoding\facebookconnector\services\EntryPersist as EntryPersistService;
 use itscoding\facebookconnector\services\EntryPoster as EntryPosterService;
-use itscoding\facebookconnector\services\EventFetcher as EventFetcherService;
+use itscoding\facebookconnector\services\EntryFetcher as EntryFetcherService;
 use itscoding\facebookconnector\services\TokenLoader as TokenLoaderService;
 use itscoding\facebookconnector\services\EntryPoster;
-use itscoding\facebookconnector\services\EventFetcher;
+use itscoding\facebookconnector\services\EntryFetcher;
 use itscoding\facebookconnector\models\Settings;
 use itscoding\facebookconnector\services\TokenLoader;
+use itscoding\facebookconnector\variables\EntryVariable;
 use itscoding\facebookconnector\widgets\OAuth;
 use Craft;
 use craft\base\Plugin;
@@ -34,7 +37,7 @@ use yii\base\Event;
  * @since     0.1.0
  *
  * @property  EntryPosterService $entryPoster
- * @property  EventFetcherService $eventFetcher
+ * @property  EntryFetcherService $entryFetcher
  * @property  TokenLoaderService $tokenLoader
  * @property  Settings $settings
  * @method    Settings getSettings()
@@ -54,7 +57,6 @@ class FacebookConnector extends Plugin
      */
     public function init()
     {
-
         parent::init();
         self::$plugin = $this;
         // Add in our console commands
@@ -62,39 +64,41 @@ class FacebookConnector extends Plugin
             $this->controllerNamespace = 'itscoding\facebookconnector\console\controllers';
         }
         $this->setComponents([
-            'tokenLoader' => TokenLoader::class,
-            'entryPoster' => EntryPoster::class,
-            'eventFetcher' => EventFetcher::class
+            'tokenLoader' => TokenLoaderService::class,
+            'entryPoster' => EntryPosterService::class,
+            'entryFetcher' => EntryFetcherService::class,
+            'entryPersist' => EntryPersistService::class
         ]);
 
         Event::on(
-            Entry::class,
-            Entry::EVENT_AFTER_SAVE,
-            function (ModelEvent $event) {
-                if ($event->isValid) {
-                    FacebookConnector::$plugin->entryPoster->post($event->sender);
-                }
+            CraftVariable::class,
+            CraftVariable::EVENT_INIT,
+            function (Event $event) {
+                /** @var CraftVariable $variable */
+                $variable = $event->sender;
+                $variable->set('facebook', EntryVariable::class);
             }
         );
 
-        Event::on(
-            Dashboard::class,
-            Dashboard::EVENT_REGISTER_WIDGET_TYPES,
-            function (RegisterComponentTypesEvent $event) {
-                if (Craft::$app->request->get('code')) {
-                    //handle facebook callback
-                    FacebookConnector::$plugin->tokenLoader->handleCallback();
-                    $errors = FacebookConnector::$plugin->tokenLoader->getErrorMessages();
-                    if (!count($errors) > 0) {
-                        Craft::$app->session->setNotice(
-                            Craft::t('facebook-connector', 'Loaded a Valid Token')
-                        );
-                    }
-                    Craft::$app->session->setError(implode(' ', $errors));
-                }
-                $event->types[] = OAuth::class;
+        Event::on(Entry::class, Entry::EVENT_AFTER_SAVE, function (ModelEvent $event) {
+            if ($event->isValid) {
+                FacebookConnector::$plugin->entryPoster->post($event->sender);
             }
-        );
+        });
+
+        Event::on(Dashboard::class, Dashboard::EVENT_REGISTER_WIDGET_TYPES, function (RegisterComponentTypesEvent $event) {
+            if (Craft::$app->request->get('code')) {
+                FacebookConnector::$plugin->tokenLoader->handleCallback();
+                $errors = FacebookConnector::$plugin->tokenLoader->getErrorMessages();
+                if ((!count($errors)) > 0) {
+                    Craft::$app->session->setNotice(
+                        Craft::t('facebook-connector', 'Loaded a Valid Token')
+                    );
+                }
+                Craft::$app->session->setError(implode(' ', $errors));
+            }
+            $event->types[] = OAuth::class;
+        });
 
         /**
          *Todo remove this comment and code if its not needed anymore