+++ src/services/TokenLoader.php
@@ -45,21 +45,22 @@ class TokenLoader extends Component
 
     private $apiVersion = '';
 
-    private $initialized = false;
-
     /**
-     * initialize all the options
+     * this function is invoked by craft (use it like a constuctor)
      */
-    public function initialize()
+    public function init()
     {
-        $this->initialized = true;
+        parent::init();
         $settings = FacebookConnector::getInstance()->getSettings();
         $this->appId = $settings->appId;
         $this->appSecret = $settings->appSecret;
         $this->apiVersion = $settings->apiVersion;
         $this->pageId = $settings->pageId;
-        $this->baseUrl = Craft::$app->request->getHostInfo();
+        if (!$this->isCliMode()) {
+            $this->baseUrl = Craft::$app->request->getHostInfo();
+        }
     }
+
 
     /**
      * FacebookConnector::$plugin->tokenLoader->getFacebookInstance()
@@ -67,15 +68,15 @@ class TokenLoader extends Component
      */
     public function getFacebookInstance()
     {
-        if (!$this->initialized) {
-            $this->initialize();
-        }
-        return new Facebook([
+        $settings = [
             'app_id' => $this->appId,
             'app_secret' => $this->appSecret,
             'default_graph_version' => $this->apiVersion,
-            'persistent_data_handler' => 'session'
-        ]);
+        ];
+        if (!$this->isCliMode()) {
+            $settings['persistent_data_handler'] = 'session';
+        }
+        return new Facebook($settings);
     }
 
     /**
@@ -240,5 +241,13 @@ class TokenLoader extends Component
     public function getErrorMessages(): array
     {
         return $this->errorMessages;
+    }
+
+    /**
+     * @return bool
+     */
+    public function isCliMode()
+    {
+        return php_sapi_name() === 'cli';
     }
 }