+++ CHANGELOG.md
@@ -4,10 +4,23 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## v0.18.0 - 2017-11-05
+### Added
+   - Entries can now be Loaded from Facebook
+   - Added a craft variable to get the entries from the database
+   - Entries now "fechtable" with quries (like standard craft contents)
+    
+## v0.17.0 - 2017-11-02
+   - Started with Implementation of "fetch Facebook Feed" 
+
+#### Pre Released 
+____
 ## v0.16.0 - 2017-10-08
 ### Added
   - Contribute Text
   - Added Link to the Youtube Configuration description
+### Fixed
+  - Updated the Documentation how its possible to use functions in `fieldconfig.php` #13
   
 ### Changed
   - Refactored the EntryPoster